requirements.txt
langchain==0.3.4
langchain_community==0.3.3
python-docx==1.1.2
chromadb==0.5.15
pandas==2.2.3
PyPDF2==3.0.1
tabula-py==2.10.0
jpype1==1.5.0
